Split PSDFormationDeducer
This simplifies introducing mx
